+++ SwiftMessages/PassthroughWindow.swift
@@ -9,6 +9,7 @@
 import UIKit
 
 class PassthroughWindow: UIWindow {
+
     override func hitTest(_ point: CGPoint, with event: UIEvent?) -> UIView? {
         let view = super.hitTest(point, with: event)
         return view == self ? nil : view