+++ SwiftMessages/Presenter.swift
@@ -10,7 +10,9 @@ import UIKit
 
 class Weak<T: AnyObject> {
     weak var value : T?
-    init() {}
+    init(value: T?) {
+        self.value = value
+    }
 }
 
 protocol PresenterDelegate: class {
@@ -21,11 +23,34 @@ protocol PresenterDelegate: class {
 
 class Presenter: NSObject, UIGestureRecognizerDelegate {
 
+    enum PresentationContext {
+        case viewController(_: Weak<UIViewController>)
+        case view(_: Weak<UIView>)
+        
+        func viewControllerValue() -> UIViewController? {
+            switch self {
+            case .viewController(let weak):
+                return weak.value
+            case .view:
+                return nil
+            }
+        }
+        
+        func viewValue() -> UIView? {
+            switch self {
+            case .viewController(let weak):
+                return weak.value?.view
+            case .view(let weak):
+                return weak.value
+            }
+        }
+    }
+    
     let config: SwiftMessages.Config
     let view: UIView
     weak var delegate: PresenterDelegate?
     let maskingView = PassthroughView()
-    let presentationContext = Weak<UIViewController>()
+    var presentationContext = PresentationContext.viewController(Weak<UIViewController>(value: nil))
     let panRecognizer: UIPanGestureRecognizer
     var translationConstraint: NSLayoutConstraint! = nil
     
@@ -49,22 +74,47 @@ class Presenter: NSObject, UIGestureRecognizerDelegate {
         let duration: TimeInterval?
         switch self.config.duration {
         case .automatic:
-            duration = 2.0
+            duration = 2
         case .seconds(let seconds):
             duration = seconds
-        case .forever:
+        case .forever, .indefinite:
             duration = nil
         }
         return duration
     }
-    
+
+    var showDate: Date?
+
+    private var interactivelyHidden = false;
+
+    var delayShow: TimeInterval? {
+        if case .indefinite(let opts) = config.duration { return opts.delay }
+        return nil
+    }
+
+    /// Returns the required delay for hiding based on time shown
+    var delayHide: TimeInterval? {
+        if interactivelyHidden { return 0 }
+        if case .indefinite(let opts) = config.duration, let showDate = showDate {
+            let timeIntervalShown = -showDate.timeIntervalSinceNow
+            return max(0, opts.minimum - timeIntervalShown)
+        }
+        return nil
+    }
+
     func show(completion: @escaping (_ completed: Bool) -> Void) throws {
-        try presentationContext.value = getPresentationContext()
+        try presentationContext = getPresentationContext()
         install()
-        showAnimation(completion: completion)
+        self.config.eventListeners.forEach { $0(.willShow) }
+        showAnimation() { completed in
+            completion(completed)
+            if completed {
+                self.config.eventListeners.forEach { $0(.didShow) }
+            }
+        }
     }
     
-    func getPresentationContext() throws -> UIViewController {
+    func getPresentationContext() throws -> PresentationContext {
         
         func newWindowViewController(_ windowLevel: UIWindowLevel) -> UIViewController {
             let viewController = WindowViewController(windowLevel: windowLevel, config: config)
@@ -74,14 +124,19 @@ class Presenter: NSObject, UIGestureRecognizerDelegate {
         switch config.presentationContext {
         case .automatic:
             if let rootViewController = UIApplication.shared.keyWindow?.rootViewController {
-                return rootViewController.sm_selectPresentationContextTopDown(config)
+                let viewController = rootViewController.sm_selectPresentationContextTopDown(config)
+                return .viewController(Weak(value: viewController))
             } else {
                 throw SwiftMessagesError.noRootViewController
             }
         case .window(let level):
-            return newWindowViewController(level)
+            let viewController = newWindowViewController(level)
+            return .viewController(Weak(value: viewController))
         case .viewController(let viewController):
-            return viewController.sm_selectPresentationContextBottomUp(config)
+            let viewController = viewController.sm_selectPresentationContextBottomUp(config)
+            return .viewController(Weak(value: viewController))
+        case .view(let view):
+            return .view(Weak(value: view))
         }
     }
     
@@ -90,28 +145,29 @@ class Presenter: NSObject, UIGestureRecognizerDelegate {
      */
     
     func install() {
-        guard let presentationContext = presentationContext.value else { return }
-        if let windowViewController = presentationContext as? WindowViewController {
-            windowViewController.install()
+        guard let containerView = presentationContext.viewValue() else { return }
+        if let windowViewController = presentationContext.viewControllerValue() as? WindowViewController {
+            windowViewController.install(becomeKey: becomeKeyWindow)
         }
-        let containerView: UIView = presentationContext.view
         do {
             maskingView.translatesAutoresizingMaskIntoConstraints = false
-            if let nav = presentationContext as? UINavigationController {
+            if let nav = presentationContext.viewControllerValue() as? UINavigationController {
                 containerView.insertSubview(maskingView, belowSubview: nav.navigationBar)
-            } else if let tab = presentationContext as? UITabBarController {
+            } else if let tab = presentationContext.viewControllerValue() as? UITabBarController {
                 containerView.insertSubview(maskingView, belowSubview: tab.tabBar)
             } else {
                 containerView.addSubview(maskingView)
             }
             let leading = NSLayoutConstraint(item: maskingView, attribute: .leading, relatedBy: .equal, toItem: containerView, attribute: .leading, multiplier: 1.00, constant: 0.0)
             let trailing = NSLayoutConstraint(item: maskingView, attribute: .trailing, relatedBy: .equal, toItem: containerView, attribute: .trailing, multiplier: 1.00, constant: 0.0)
-            let top = topLayoutConstraint(view: maskingView, presentationContext: presentationContext)
-            let bottom = bottomLayoutConstraint(view: maskingView, presentationContext: presentationContext)
+            let top = topLayoutConstraint(view: maskingView, containerView: containerView, viewController: presentationContext.viewControllerValue())
+            let bottom = bottomLayoutConstraint(view: maskingView, containerView: containerView, viewController: presentationContext.viewControllerValue())
             containerView.addConstraints([top, leading, bottom, trailing])
         }
         do {
             view.translatesAutoresizingMaskIntoConstraints = false
+            // Have the container layout now so we can see exactly where `maskingView`
+            containerView.layoutIfNeeded()
             maskingView.addSubview(view)
             let leading = NSLayoutConstraint(item: view, attribute: .leading, relatedBy: .equal, toItem: maskingView, attribute: .leading, multiplier: 1.00, constant: 0.0)
             let trailing = NSLayoutConstraint(item: view, attribute: .trailing, relatedBy: .equal, toItem: maskingView, attribute: .trailing, multiplier: 1.00, constant: 0.0)
@@ -129,15 +185,13 @@ class Presenter: NSObject, UIGestureRecognizerDelegate {
                 case .top:
                     top += adjustable.bounceAnimationOffset
                     if !UIApplication.shared.isStatusBarHidden {
-                        if let vc = presentationContext as? WindowViewController {
+                        if let vc = presentationContext.viewControllerValue() as? WindowViewController {
                             if vc.windowLevel == UIWindowLevelNormal {
                                 top += adjustable.statusBarOffset
                             }
-                        } else if let vc = presentationContext as? UINavigationController {
-                            if !vc.sm_isVisible(view: vc.navigationBar) {
-                                top += adjustable.statusBarOffset
-                            }
-                        } else {
+                        } else if let vc = presentationContext.viewControllerValue() as? UINavigationController, !vc.sm_isVisible(view: vc.navigationBar) {
+                            top += adjustable.statusBarOffset
+                        } else if viewInterferesWithStatusBar(maskingView) {
                             top += adjustable.statusBarOffset
                         }
                     }
@@ -159,15 +213,16 @@ class Presenter: NSObject, UIGestureRecognizerDelegate {
                 if interactive {
                     maskingView.tappedHander = { [weak self] in
                         guard let strongSelf = self else { return }
-                        self?.delegate?.hide(presenter: strongSelf)
+                        strongSelf.interactivelyHidden = true
+                        strongSelf.delegate?.hide(presenter: strongSelf)
                     }
                 } else {
                     // There's no action to take, but the presence of
                     // a tap handler prevents interaction with underlying views.
                     maskingView.tappedHander = { }
                 }
+                maskingView.accessibilityViewIsModal = true
             }
-            
             switch config.dimMode {
             case .none:
                 break
@@ -178,19 +233,39 @@ class Presenter: NSObject, UIGestureRecognizerDelegate {
             }
         }
     }
-    
-    func topLayoutConstraint(view: UIView, presentationContext: UIViewController) -> NSLayoutConstraint {
-        if case .top = config.presentationStyle, let nav = presentationContext as? UINavigationController, nav.sm_isVisible(view: nav.navigationBar) {
-            return NSLayoutConstraint(item: view, attribute: .top, relatedBy: .equal, toItem: nav.navigationBar, attribute: .bottom, multiplier: 1.00, constant: 0.0)
+
+    private var becomeKeyWindow: Bool {
+        if config.becomeKeyWindow == .some(true) { return true }
+        switch config.dimMode {
+        case .gray, .color:
+            // Should become key window in modal presentation style
+            // for proper voice over handling.
+            return true
+        case .none:
+            return false
         }
-        return NSLayoutConstraint(item: view, attribute: .top, relatedBy: .equal, toItem: presentationContext.view, attribute: .top, multiplier: 1.00, constant: 0.0)
     }
 
-    func bottomLayoutConstraint(view: UIView, presentationContext: UIViewController) -> NSLayoutConstraint {
-        if case .bottom = config.presentationStyle, let tab = presentationContext as? UITabBarController, tab.sm_isVisible(view: tab.tabBar) {
+    private func viewInterferesWithStatusBar(_ view: UIView) -> Bool {
+        guard let window = view.window else { return false }
+        let statusBarFrame = UIApplication.shared.statusBarFrame
+        let statusBarWindowFrame = window.convert(statusBarFrame, from: nil)
+        let statusBarViewFrame = view.convert(statusBarWindowFrame, from: nil)
+        return statusBarViewFrame.intersects(view.bounds)
+    }
+    
+    func topLayoutConstraint(view: UIView, containerView: UIView, viewController: UIViewController?) -> NSLayoutConstraint {
+        if case .top = config.presentationStyle, let nav = viewController as? UINavigationController, nav.sm_isVisible(view: nav.navigationBar) {
+            return NSLayoutConstraint(item: view, attribute: .top, relatedBy: .equal, toItem: nav.navigationBar, attribute: .bottom, multiplier: 1.00, constant: 0.0)
+        }
+        return NSLayoutConstraint(item: view, attribute: .top, relatedBy: .equal, toItem: containerView, attribute: .top, multiplier: 1.00, constant: 0.0)
+    }
+
+    func bottomLayoutConstraint(view: UIView, containerView: UIView, viewController: UIViewController?) -> NSLayoutConstraint {
+        if case .bottom = config.presentationStyle, let tab = viewController as? UITabBarController, tab.sm_isVisible(view: tab.tabBar) {
             return NSLayoutConstraint(item: view, attribute: .bottom, relatedBy: .equal, toItem: tab.tabBar, attribute: .top, multiplier: 1.00, constant: 0.0)
         }
-        return NSLayoutConstraint(item: view, attribute: .bottom, relatedBy: .equal, toItem: presentationContext.view, attribute: .bottom, multiplier: 1.00, constant: 0.0)
+        return NSLayoutConstraint(item: view, attribute: .bottom, relatedBy: .equal, toItem: containerView, attribute: .bottom, multiplier: 1.00, constant: 0.0)
     }
     
     /*
@@ -219,7 +294,6 @@ class Presenter: NSObject, UIGestureRecognizerDelegate {
     }
 
     func showViewAnimation(completion: @escaping (_ completed: Bool) -> Void) {
-        
         switch config.presentationStyle {
         case .top, .bottom:
             let animationDistance = self.translationConstraint.constant + bounceOffset
@@ -231,11 +305,15 @@ class Presenter: NSObject, UIGestureRecognizerDelegate {
                 self.view.superview?.layoutIfNeeded()
                 }, completion: { completed in
                     completion(completed)
-            })
+                })
         }
     }
 
+    var isHiding = false
+
     func hide(completion: @escaping (_ completed: Bool) -> Void) {
+        isHiding = true
+        self.config.eventListeners.forEach { $0(.willHide) }
         switch config.presentationStyle {
         case .top, .bottom:
             UIView.animate(withDuration: 0.2, delay: 0, options: [.beginFromCurrentState, .curveEaseIn], animations: {
@@ -243,11 +321,12 @@ class Presenter: NSObject, UIGestureRecognizerDelegate {
                 self.translationConstraint.constant -= size.height
                 self.view.superview?.layoutIfNeeded()
                 }, completion: { completed in
-                    if let viewController = self.presentationContext.value as? WindowViewController {
+                    if let viewController = self.presentationContext.viewControllerValue() as? WindowViewController {
                         viewController.uninstall()
                     }
                     self.maskingView.removeFromSuperview()
-                    completion(completed)
+                    completion(true)
+                    self.config.eventListeners.forEach { $0(.didHide) }
             })
 // TODO the spring animation makes the interactive hide transition smoother, but
 // TODO the added delay due to damping makes status bar style transitions look bad.
@@ -267,7 +346,7 @@ class Presenter: NSObject, UIGestureRecognizerDelegate {
 //                completion(completed: completed)
 //            })
         }
-        
+
         func undim() {
             UIView.animate(withDuration: 0.2, animations: {
                 self.maskingView.backgroundColor = UIColor.clear
@@ -329,6 +408,7 @@ class Presenter: NSObject, UIGestureRecognizerDelegate {
             panTranslationY = translation.y
         case .ended, .cancelled:
             if closeSpeed > 750.0 || closePercent > 0.33 {
+                interactivelyHidden = true
                 delegate?.hide(presenter: self)
             } else {
                 closing = false
@@ -369,3 +449,4 @@ class Presenter: NSObject, UIGestureRecognizerDelegate {
         return true
     }
 }
+